requirements.txt
pyTelegramBotAPI
yt-dlp
google-generativeai
urlextract
